created screen for note viewing with generic implementation for viewing notes of a notebook

+++ Android/app/src/main/java/com/fivegearszerochill/noted/repository/NoteRepository.java
@@ -35,7 +35,7 @@ public class NoteRepository {
                 .build();
     }
 
-    public LiveData<PagedList<NoteEntity>> getNotesAsync(int notebookId) {
+    public LiveData<PagedList<NoteEntity>> getNotesAsync(long notebookId) {
         DataSource.Factory<Integer, NoteEntity> dataSource = noteDao.getNotesPaginated(notebookId);
         return new LivePagedListBuilder<>(
                 dataSource,
